sentryflow/requirements.txt
# Core Dependencies
streamlit==1.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0
altair==5.2.0

# Caching & Performance
redis==5.0.1
cachetools==5.3.2

# AI/ML (for more sophisticated checks)
transformers==4.37.0
sentence-transformers==2.3.1
torch==2.1.2

# Utilities
pyyaml==6.0.1
jsonschema==4.20.0
python-dateutil==2.8.2

# Testing (optional for hackathon)
pytest==8.0.0
pytest-cov==4.1.0
